requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.2
Django==4.2.8
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-profanity-check==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
inflection==0.5.1
joblib==1.3.2
numpy==1.26.2
packaging==23.2
pillow==10.2.0
profanity-check==1.0.3
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
scipy==1.11.4
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.9.0
whitenoise==6.6.0
